+++ lib/Config.ts
@@ -150,36 +150,4 @@ class Config {
   }
 }
 
-export const convertConfigToToml = () => {
-  let result = '';
-
-  const recursivelyConvertJsonToToml = (json: any, prefix: string) => {
-    const nestedPairs: any = [];
-    const simplePairs: any = [];
-
-    if (!json) return;
-
-    Object.keys(json).sort().forEach((key: any) => {
-      if (Object.prototype.toString.call(json[key]) !== '[object Function]') {
-        (isPlainObject(json[key]) ? nestedPairs : simplePairs).push([key, json[key]]);
-      }
-    });
-
-    if (!(prefix === '' || simplePairs.length === 0)) {
-      result += '[' + prefix + ']\n';
-    }
-
-    simplePairs.forEach((value: any) => {
-      result += value[0] + ' = ' + value[1] + '\n';
-    });
-    nestedPairs.forEach((value: any) => {
-      recursivelyConvertJsonToToml(value[1], (prefix === null || prefix === '') ? value[0] : [prefix, value[0]].join('.'));
-    });
-  };
-
-  recursivelyConvertJsonToToml(new Config(), '');
-
-  fs.writeFileSync('../xud.conf', result);
-};
-
 export default Config;
