+++ tasks/config/index.ts
@@ -1,0 +1,1 @@
+export { default as sample } from './sample';
