+++ tasks/db/dropTables.ts
@@ -1,3 +1,4 @@
+/*
 import Config from '../../lib/Config';
 import DB from '../../lib/db/DB';
 import Logger from '../../lib/Logger';
@@ -14,3 +15,4 @@ export default async (testDb?: boolean) => {
   await db.dropTables();
   db.close();
 };
+*/
