+++ tasks/db/index.ts
@@ -1,4 +1,6 @@
+/*
 export { default as dropTables } from './dropTables';
 export { default as init } from './init';
 export { default as restart } from './restart';
 export { default as truncate } from './truncate';
+*/
