+++ tasks/db/restart.ts
@@ -1,3 +1,4 @@
+/*
 import dropTables from './dropTables';
 import init from './init';
 
@@ -5,3 +6,4 @@ export default async (testDb?: boolean) => {
   await dropTables(testDb);
   await init(testDb);
 };
+*/
