+++ tasks/index.ts
@@ -1,5 +1,7 @@
-import * as db from './db';
+// import * as db from './db';
+import * as config from './config';
 
 export default {
-  db,
+  // db,
+  config,
 };
